back/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Faker==35.0.0
faker-vehicle==0.2.0
fonttools==4.56.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
lxml==5.3.0
mysqlclient==2.2.7
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pypdf==5.1.0
pyphen==0.17.2
pytest==8.3.5
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
